Fixed warnings when building the bfs_shell.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@23462 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/file_systems/bfs/BPlusTree.cpp
@@ -1002,7 +1002,7 @@ BPlusTree::_SplitNode(bplustree_node *node, off_t nodeOffset,
 	int32 keyIndex = *_keyIndex;	// can become less than zero!
 
 	if (keyIndex > node->NumKeys()) {
-		FATAL(("key index out of bounds: %ld, num keys: %u\n", keyIndex,
+		FATAL(("key index out of bounds: %d, num keys: %u\n", (int)keyIndex,
 			node->NumKeys()));
 		return B_BAD_VALUE;
 	}
